the tests below still have runtime errors but the others work fine // jpeg // tiff // gs // ispell // bf // pgp // rawaudio

+++ mibench/consumer/tiff-v3.5.4/libtiff/tif_dirread.c
@@ -127,7 +127,8 @@ TIFFReadDirectory(TIFF* tif)
 		 */
 		(void) ReadOK(tif, &nextdiroff, sizeof (uint32));
 	} else {
-		toff_t off = tif->tif_diroff;
+		//fixed : the casting was needed in new platforms
+		toff_t off = (int)tif->tif_diroff;
 
 		if ((tsize_t) (off + sizeof (uint16)) > tif->tif_size) {
 			TIFFError(tif->tif_name,
